app/requirements.txt
Flask==2.2.5
mysql-connector-python==8.2.0
